+++ build_cli_annotations/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 2.1.1-wip
 
 - Require `args: ^2.6.0`
-- Require `sdk: ^3.5.0`
+- Require `sdk: ^3.6.0`
 
 ## 2.1.0